refactor: move utils to notion dir
